+++ _src/managers/providers/browser.ts
@@ -37,7 +37,7 @@ export class NaniumBrowserProviderConfig {
 
 export class NaniumProviderBrowser implements ServiceProviderManager {
 	repository: NaniumRepository;
-	internalEventSubscriptions: { [eventName: string]: ((event: any) => void)[] } = {};
+	internalEventSubscriptions: { [eventName: string]: EventSubscription[] } = {};
 	config: NaniumBrowserProviderConfig = {
 		isResponsible: async (): Promise<number> => Promise.resolve(1),
 		isResponsibleForEvent: async (): Promise<number> => Promise.resolve(1),
@@ -161,8 +161,8 @@ export class NaniumProviderBrowser implements ServiceProviderManager {
 
 	async emit(eventName: string, event: any, executionContext: ExecutionContext): Promise<void> {
 		if (this.internalEventSubscriptions[eventName]?.length) {
-			for (const handler of this.internalEventSubscriptions[eventName]) {
-				handler(event);
+			for (const s of this.internalEventSubscriptions[eventName]) {
+				s.handler(event);
 			}
 		}
 	}
@@ -173,16 +173,19 @@ export class NaniumProviderBrowser implements ServiceProviderManager {
 
 	async subscribe(eventConstructor: new() => any, handler: EventHandler, context?: ExecutionContext): Promise<EventSubscription> {
 		const eventName: string = (eventConstructor as any).eventName;
-		this.internalEventSubscriptions[eventName] =
-			this.internalEventSubscriptions[eventName] ?? [];
-		this.internalEventSubscriptions[eventName].push(handler);
-		return new EventSubscription('', eventName, handler);
+		const subscription = new EventSubscription('', eventName, handler);
+		this.internalEventSubscriptions[eventName] = this.internalEventSubscriptions[eventName] ?? [];
+		this.internalEventSubscriptions[eventName].push(subscription);
+		return subscription;
 	}
 
-	async unsubscribe(eventConstructor: any, handler?: (data: any) => Promise<void>): Promise<void> {
-		const eventName: string = eventConstructor.eventName;
-		if (handler) {
-			this.internalEventSubscriptions[eventName] = this.internalEventSubscriptions[eventName].filter(h => h !== handler);
+	async unsubscribe(subscription?: EventSubscription, eventName?: string): Promise<void> {
+		eventName = subscription?.eventName ?? eventName;
+		if (subscription) {
+			if (eventName in this.internalEventSubscriptions) {
+				this.internalEventSubscriptions[eventName] = this.internalEventSubscriptions[eventName]
+					.filter(s => s !== subscription);
+			}
 		} else {
 			delete this.internalEventSubscriptions[eventName];
 		}