+++ _src/tests/events/eventBase.ts
@@ -3,8 +3,10 @@ import { Nanium } from '../../core';
 import { EventHandler } from '../../interfaces/eventHandler';
 import { EventSubscription } from '../../interfaces/eventSubscription';
 import { ServiceManager } from '../../interfaces/serviceManager';
+import { NaniumObject } from '../../objects';
 
-export class EventBase {
+export class EventBase<T = any> extends NaniumObject<T> {
+
 	emit(context?: ExecutionContext): void {
 		Nanium.emit(this, undefined, context);
 	}
@@ -13,10 +15,11 @@ export class EventBase {
 		return await Nanium.subscribe(this, handler, managerOrData);
 	}
 
-	static async unsubscribe(subscription: EventSubscription): Promise<void> {
-		await Nanium.unsubscribe(subscription);
+	static async unsubscribe(subscription?: EventSubscription): Promise<void> {
+		await Nanium.unsubscribe(subscription, this['eventName']);
 	}
 }
+
 
 export class TestSubscriptionData {
 	constructor(