+++ _src/tests/events/test/stuff2Event.ts
@@ -1,0 +1,9 @@
+import { EventBase } from '../eventBase';
+import { EventType } from '../../../objects';
+
+@EventType({
+	scope: 'public'
+})
+export class Stuff2Event extends EventBase<Stuff2Event> {
+	static eventName: string = 'NaniumTest:test/stuff2';
+}