+++ _src/tests/services/test/get.executor.ts
@@ -3,6 +3,7 @@ import { TestGetRequest, TestGetResponse } from './get.contract';
 import { ServiceExecutor } from '../../../interfaces/serviceExecutor';
 import { StuffEvent } from '../../events/test/stuffEvent';
 import { TestExecutionContext } from '../testExecutionContext';
+import { Stuff2Event } from '../../events/test/stuff2Event';
 
 export class TestGetExecutor implements ServiceExecutor<TestGetRequest, TestGetResponse> {
 	static serviceName: string = 'NaniumTest:test/get';
@@ -18,10 +19,12 @@ export class TestGetExecutor implements ServiceExecutor<TestGetRequest, TestGetR
 			throw new Error('no no!');
 		}
 		new StuffEvent(9, '10', new Date(2011, 11, 11)).emit(executionContext);
+		new Stuff2Event().emit(executionContext);
 		return new TestGetResponse({
 			output1: request.body.input1 + ' :-)',
 			output2: 2
 		});
 	}
 }
+